Rewrite debian/rules using dh driver
Bang on the debian/rules file until it works.  Changes in behavior:

 - the test suite was not being run.

 - the DEB_BUILD_OPTIONS settings were not being respected.
   Now nocheck is respected and at least we complain about noopt
   (the optimization level is not configurable in the upstream
   makefile).

 - the ‘debian/rules clean’ rule was flaky.

 - the GOROOT dir was hard-coded to /usr/lib/google-go.  Now it
   is /usr/lib/golang to match the upstream domain name, but more
   importantly it is configurable.

 - there were a bunch of hardcoded wrappers in the debian/bin
   dir, which made for unpleasant editing.  Now they are automatically
   built as needed.

 - the wrappers did not let the user override the GO* parameters
   through her own environment.

 - files needed to run "godoc -http=:6060" were not installed.

 - the rules for determining GOARCH, GOARM, etc were not cross-
   compilation friendly.

Since I had to rewrite debian/rules anyway, I went ahead and put the
result in the public domain so it can be reused by upstream and other
packages.

Now it works for me as far as "godoc -http=:6060" on my 32-bit x86; no
guarantees beyond that, though.  Please break it so we can get
something robust ready for release.