compile golang-go-linux-* for each arch
Before this commit, golang-go-* were Architecture: all, meaning that
they are built only once on — typically — either an amd64 or i386
machine.

However, cgo is only available when building natively, meaning that when
building on amd64, only golang-go-linux-amd64 had cgo enabled, or when
building on i386, only golang-go-linux-386 had cgo enabled.

Given that the packages behave differently depending on which
architecture they are compiled, “Architecture: all” is obviously wrong.

For golang-go-*-* where * is !linux, this is not a concern, since these
packages can _only_ be used when cross-compiling anyway. When golang-go
becomes available on more architectures, this needs to be changed.